add header

+++ 00/libft/ft_calloc.c
@@ -1,9 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_calloc.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: riamaev <riamaev@student.hive.fi>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/10/29 09:29:08 by riamaev           #+#    #+#             */
+/*   Updated: 2024/10/29 12:36:28 by riamaev          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-void	*ft_calloc(int nmem, int size)
+void	*ft_calloc(size_t nmem, size_t size)
 {
-	int		total;
-	void	*ptr;
+	size_t		total;
+	void		*ptr;
 
 	total = nmem * size;
 	if (nmem != 0 && total / nmem != size)
@@ -19,10 +31,10 @@ void	*ft_calloc(int nmem, int size)
 
 int	main(void)
 {
-	int *arr;
-	int n = 10;
-	int i = 0;
-	arr = (int *)ft_calloc(n, sizeof(int));
+	size_t *arr;
+	size_t n = 10;
+	size_t i = 0;
+	arr = (size_t *)ft_calloc(n, sizeof(size_t));
 	if (arr == NULL)
 	{
 		printf("memory allocation failed\n");
@@ -30,7 +42,7 @@ int	main(void)
 	}
 	while (i < n)
 	{
-		printf("%d\n", arr[i++]);
+		printf("%ld\n", arr[i++]);
 	}
 	free (arr);
 	return (0);
